seperated rainbow functionality from runner script
